rename components directory Operator -> Staking
